Add files via upload
Q1 : asking location wasn't there when I use Chrome Dev Tool, but showed when run thy .py file. I don't know how to handle it. 
Account & Lists and Orders are same CSS Selcetor.. How to distinguish it? 
Q2 : Step failed but I don't know how to figure it out and fix my code??
And I moved img file to git folder and add it. and saw it for git status. but it wasn't git push... so I uploaded it on github.com